Split Beam example from Analyzer example.
Change limited beam boundaries to sys.maxint
